Fix warnings about unescaped { on perl 5.17.6+
Unforunately, when you have

```perl
$x =~ s{ \{  }{ bar  }
```

All that the \ escapes is "is this a regexp token", ie:

is no different to its behaviour inside

```perl
$x =~ s[ \{ ][ bar ]
```

And upstream changes mean you now have to escape the delimiting as well now, so I think

```perl
$x =~ s{ \\{  }{ }
```

Is the "right" way to do it. Though, I'm not entirely sure about that either.

Either way, its easier to chose a non-conflicting delimiter token

```perl
$x =~ s[ \{ ][ ]
```

and that should do what you want.